[fix] viewCount, answerCount Long -> int로 변경 및 count Default 0 설정

+++ back/stackoverflow/src/main/java/com/teambj/stackoverflow/domain/question/service/QuestionService.java
@@ -67,7 +67,7 @@ public class QuestionService {
         return questionRepository.findAll();
     }
 
-    public void updateQuestionViewCount(Question question, Long viewCount) {
+    public void updateQuestionViewCount(Question question, int viewCount) {
         question.setViewCount(viewCount + 1);
         questionRepository.save(question);
     }
@@ -77,13 +77,10 @@ public class QuestionService {
         questionRepository.delete(question);
     }
 
-
     private Question findVerifiedQuestionById(Long questionId) {
         Optional<Question> optionalQuestion = questionRepository.findById(questionId);
         Question foundQuestion = optionalQuestion.orElseThrow(() -> new BusinessLogicException(ExceptionCode.QUESTION_NOT_FOUND));
 
         return foundQuestion;
     }
-
-
 }
